+++ CHANGELOG.md
@@ -4,7 +4,6 @@
 ### Breaking Changes
 #### App
 - Removed `App:InterProcessCommunicator` (can be easily recreated using the `Network::Socket` class)
-- Added `isPortable` parameter to `DataFileManager` to store config files in the executable directory instead
 #### Logging
 - Removed module
 #### Network
@@ -12,7 +11,14 @@
 - Renamed `Web::fetchJson()` to `Web::getJson()`
 - Removed `CurlEasy` class
 ### New APIs
-None
+#### App
+- Added `isPortable` parameter to `DataFileManager` to store config files in the executable directory instead
+#### Helpers
+- Added `PairHash` class
+- Added `CodeHelpers::combineHash()` function
+#### System
+- Added `DependencySearchOption` enum
+- Added `search` parameter to `Environment::findDependency()` to specify a `DependencySearchOption`
 ### Fixes
 None